Simplified handling of LICENSE.TXT.
For some reason, LICENSE.TXT was specified in debian/docs,
which caused a lintian warning to be raised. Instead of
removing the reference from debian/docs, the file was
excluded from installation in override_dh_installdocs,
which is just redundant.

Removed both references so LICENSE.TXT is not installed
or removed. This is better and also lintian-clean.

Thanks to Luciano Bello for noticing this.