Compare packet header sizes as signed
This commit fixes -Wsign-compare warnings that are raised because of a
packet capture length validation. The code being fixed does something
like this:

unsigned char *packet_end;
unsigned char *packet_start;
if (packet_end - packet_start < sizeof(...) + sizeof(...)) {
  ...
}

The left side type, being a difference of pointers is ptrdiff_t, which
is signed by definition [1].

The right side type, being sizeof() is size_t, which is usually
unsigned.

There were two options:
 a) to cast the left side as unsigned
 b) to cast the right side to signed

To be honest, I don't like either of them to be chosen universally. I
would rather verify each case and make sure that either side is within
some limits and do the cast to the appropriate data type.

However, packet lengths will most likely never exceed 32 KiB, so casting
the right side to signed integer would be a good way to go. Just in
case, I decided to cast to signed *long* integer to be on the safe side.
This should be enough for a good while.

I decided for this as it is more predictable. I did not find a trivial
way to guarantee that packet_end - packet_start would always be > 0, and
thus, castable to unsigned long integer. An alternative is to add an
assert( packet_start >= packet_end) and wait for crash reports but this
might be too intrusive.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/stddef.h.html